Connect Django project with env file
Django project doesn't run with the same database parameters on localhost and AWS machine. Therefore I created two different env files named env_sample and prod_sample. In or der to run the project on localhost a .env file should be created and parameters in env_sample should be used. To run it on AWS machine prod_sample parameters need to be used.